remove portfolio changes
